+++ package.json
@@ -14,19 +14,23 @@
     "@fullcalendar/timegrid": "^4.4.0",
     "@material-ui/core": "^4.9.7",
     "@material-ui/icons": "^4.9.1",
+    "@material-ui/lab": "^4.0.0-alpha.47",
     "@material-ui/pickers": "^3.2.10",
     "@material-ui/styles": "^4.9.6",
     "axios": "^0.19.2",
     "clsx": "^1.1.0",
     "date-fns": "^2.11.1",
     "dotenv": "^8.2.0",
+    "final-form": "^4.18.7",
     "material-ui-pickers": "^2.2.4",
+    "mui-rff": "^2.1.1",
     "node-sass": "^4.13.1",
     "popper.js": "^1.16.1",
     "prettier": "^2.0.2",
     "prop-types": "^15.7.2",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-final-form": "^6.3.5",
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.1.0",
     "react-scripts": "^3.4.1",
@@ -35,7 +39,8 @@
     "redux": "^4.0.5",
     "redux-thunk": "^2.3.0",
     "tar": "^4.4.11",
-    "tippy.js": "^4.3.5"
+    "tippy.js": "^4.3.5",
+    "yup": "^0.28.3"
   },
   "homepage": "https://omoulearning.com",
   "scripts": {
