+++ src/components/Form/Form.scss
@@ -76,3 +76,7 @@
     margin-left: 5px;
   }
 }
+
+.select {
+  width: 200px;
+}
