+++ src/reducers/authReducer.js
@@ -9,8 +9,6 @@ export default function auth(state = initialState.Authentication, {payload, type
             return onLogout(state);
         case actions.FETCH_USER_SUCCESSFUL:
             return onDetailFetch(state, payload);
-        case actions.FETCH_USER_FAILED:
-            return state;
         default:
             return state;
     }
